+++ src/type-definition/QuickPickItemExType.ts
@@ -4,6 +4,7 @@ interface ExtendedQuickPickItem extends vscode.QuickPickItem {
     value: string;
 }
 
-type QuickPickItemEx = ExtendedQuickPickItem | vscode.QuickPickItem;
+// type QuickPickItemEx = ExtendedQuickPickItem | vscode.QuickPickItem;
+type QuickPickItemEx = ExtendedQuickPickItem;
 
 export default QuickPickItemEx;